fix(combat): address review findings — null focused_action, fled completion, API safety
Critical: guard _resolve_pc_action and combo detection against null
focused_action, filter fled participants from encounter completion.
Moderate: prefetch vitals, cache is_gm, fix timer task locking, add
input serializers, restrict join to declaration/between-rounds, type
wounds_applied.
Minor: threshold type consistency, round_started_at in service layer.

+++ src/world/vitals/constants.py
@@ -26,7 +26,7 @@ class CharacterStatus(models.TextChoices):
 
 PERMANENT_WOUND_THRESHOLD: float = 0.5
 KNOCKOUT_HEALTH_THRESHOLD: float = 0.2
-DEATH_HEALTH_THRESHOLD: int = 0
+DEATH_HEALTH_THRESHOLD: float = 0.0
 
 # ---------------------------------------------------------------------------
 # Wound descriptions -- (threshold, description) from healthiest to worst
